fix error handling

+++ hamburgers/viewsets.py
@@ -3,7 +3,6 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.reverse import reverse
 from rest_framework.views import APIView
-from django.http import Http404, HttpResponseBadRequest
 
 from .serializers import HamburgerSerializer, IngredientSerializer
 from .models import Hamburger, Ingredient
@@ -20,19 +19,25 @@ class HamburgerDetail(APIView):
     """
     def get_object(self, pk):
         if not isinstance(pk, int):
-            raise HttpResponseBadRequest
+            return 'not instance'
         try:
             return Hamburger.objects.get(pk=pk)
         except Hamburger.DoesNotExist:
-            raise Http404
+            return 'not instance'
 
     def get(self, request, pk, format=None):
         hamburger = self.get_object(pk)
+        if hamburger == 'not instance':
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        if hamburger == 'doesnt exist':
+            return Response(status=status.HTTP_404_NOT_FOUND)
         serializer = HamburgerSerializer(hamburger)
         return Response(serializer.data)
 
     def patch(self, request, pk, format=None):
         hamburger = self.get_object(pk)
+        if hamburger == 'not instance' or hamburger == 'doesnt exist':
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
         poss_variables = ['nombre', 'precio' 'descripcion', 'imagen']
         for param in request.data:
@@ -51,6 +56,9 @@ class HamburgerDetail(APIView):
 
     def delete(self, request, pk, format=None):
         hamburger = self.get_object(pk)
+        if hamburger == 'not instance' or hamburger == 'doesnt exist':
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
         hamburger.delete()
         return Response(status=status.HTTP_200_OK)
 
@@ -59,16 +67,6 @@ class IngredientViewSet(viewsets.ModelViewSet):
     queryset = Ingredient.objects.all().order_by('nombre')
     serializer_class = IngredientSerializer
 
-    def destroy(self, request, *args, **kwargs):
-        try:
-            instance = self.get_object()
-            hamburgers = instance.hamburger_set.all()
-            if not hamburgers:
-                self.perform_destroy(instance)
-        except Http404:
-            pass
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
 
 class IngredientDetail(APIView):
     """
@@ -76,19 +74,27 @@ class IngredientDetail(APIView):
     """
     def get_object(self, pk):
         if not isinstance(pk, int):
-            raise HttpResponseBadRequest
+            return 'not instance'
         try:
             return Ingredient.objects.get(pk=pk)
         except Ingredient.DoesNotExist:
-            raise Http404
+            return 'doesnt exist'
 
     def get(self, request, pk, format=None):
         ingredient = self.get_object(pk)
+
+        if ingredient == 'not instance':
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        if ingredient == 'doesnt exist':
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
         serializer = IngredientSerializer(ingredient)
         return Response(serializer.data)
 
     def delete(self, request, pk, format=None):
         ingredient = self.get_object(pk)
+        if ingredient == 'not instance' or ingredient == 'doesnt exist':
+            return Response(status=status.HTTP_404_NOT_FOUND)
         hamburgers = ingredient.hamburger_set.all()
         if not hamburgers:
             ingredient.delete()
@@ -100,35 +106,44 @@ class IntermediateDetail(APIView):
 
     def get_hamburger(self, pk):
         if not isinstance(pk, int):
-            raise HttpResponseBadRequest
+            return 'not instance'
         try:
             return Hamburger.objects.get(pk=pk)
         except Hamburger.DoesNotExist:
-            raise Http404
+            return 'not instance'
 
     def get_ingredient(self, pk):
         if not isinstance(pk, int):
-            raise HttpResponseBadRequest
+            return 'doesnt exist'
         try:
             return Ingredient.objects.get(pk=pk)
         except Ingredient.DoesNotExist:
-            raise Http404
+            return 'doesnt exist'
 
     def put(self, request, pk, pk2, format=None):
         hamburger = self.get_hamburger(pk=pk)
         ingredient = self.get_ingredient(pk=pk2)
+
+        if hamburger == 'not instance':
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        if ingredient == 'doesnt exist':
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
         if ingredient not in hamburger.ingredientes.all():
             hamburger.ingredientes.add(ingredient)
         return Response(status=status.HTTP_201_CREATED)
 
     def delete(self, request, pk, pk2, format=None):
-        if not isinstance(pk, int):
-            raise HttpResponseBadRequest
-        hamburger = Hamburger.objects.get(pk=pk)
-        ingredient = Ingredient.objects.get(pk=pk2)
+        hamburger = self.get_hamburger(pk=pk)
+        ingredient = self.get_ingredient(pk=pk2)
 
-        if ingredient in hamburger.ingredients.all():
-            hamburger.ingredients.remove(ingredient)
+        if hamburger == 'not instance':
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        if ingredient == 'doesnt exist':
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if ingredient in hamburger.ingredientes.all():
+            hamburger.ingredientes.remove(ingredient)
             return Response(status=status.HTTP_200_OK)
         else:
             return Response(status=status.HTTP_404_NOT_FOUND)
